requirements.txt
absl-py==1.2.0
altgraph==0.17.3
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
contextlib2==21.6.0
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
docutils==0.19
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
filelock==3.8.0
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0.7
fonttools==4.37.1
gast==0.4.0
google==3.0.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.48.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
imutils==0.5.4
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jaraco.classes==3.2.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.10.0
keras-nightly==2.11.0.dev2022091107
Keras-Preprocessing==1.1.2
keyring==23.9.1
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
macholib==1.16.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==8.14.0
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
oauthlib==3.2.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pycparser==2.21
Pygments==2.13.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
readme-renderer==37.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
stack-data==0.5.0
tb-nightly==2.11.0a20220910
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-object-detection-api==0.1.1
termcolor==1.1.0
terminado==0.15.0
tf-estimator-nightly==2.11.0.dev2022091108
tf-nightly==2.11.0.dev20220911
tflite-support==0.1.0a1
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
twine==4.0.1
typing_extensions==4.3.0
urllib3==1.26.11
virtualenv==20.16.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
wget==3.2
widgetsnbextension==4.0.3
wrapt==1.14.1
zipp==3.8.1